+++ APBD_3/Container/GasContainer.cs
@@ -11,7 +11,7 @@ public class GasContainer(
         Console.WriteLine($"[HAZARD] container {SerialNumber}");
     }
     
-    public override void LoadCargo(int mass)
+    public override void LoadCargo(double mass)
     {
         if (mass > MaxCargoMassKg)
         {
